requirements.txt
Flask==3.0.3
pandas==2.2.3
openpyxl==3.1.5
psycopg2-binary==2.9.10
SQLAlchemy==2.0.35
gunicorn==23.0.0




